refactor: refactored code and created suggestions component
