+++ src/Pages/Home/Home.jsx
@@ -2,31 +2,22 @@ import React, { useEffect } from "react";
 import {
   Header,
   PostCard,
-  SuggestionsCard,
+  Suggestions,
   CreatePostModal,
   Sort,
 } from "../../components/";
 import { useSelector, useDispatch } from "react-redux";
-import { getUsers, getPosts } from "../../redux/asyncThunk/";
+import { getPosts } from "../../redux/asyncThunk/";
 import "./Home.css";
-import { VscNoNewline } from "react-icons/vsc";
 
 const Home = () => {
   const dispatch = useDispatch();
-  const { users } = useSelector((state) => state.users);
   const { posts } = useSelector((state) => state.posts);
   const { user } = useSelector((state) => state.auth);
 
   useEffect(() => {
-    dispatch(getUsers());
     dispatch(getPosts());
   }, [dispatch]);
-
-  const otherUsers = users.filter((item) => item.username !== user.username);
-
-  const nonFollowing = otherUsers.filter((item) =>
-    item.following.every((following) => following.username !== user.username)
-  );
 
   let feedUsers = user.followers.map((user) => user.username);
   feedUsers = [...feedUsers, user.username];
@@ -45,14 +36,7 @@ const Home = () => {
             ))}
           </section>
 
-          <section className="suggestions-container">
-            <h2 className="suggestions-title">People you may know</h2>
-            <div className="suggestions-list">
-              {nonFollowing.map((user) => (
-                <SuggestionsCard key={user._id} user={user} />
-              ))}
-            </div>
-          </section>
+          <Suggestions />
         </section>
         <CreatePostModal />
       </main>
