+++ src/Pages/Profile/Profile.jsx
@@ -1,34 +1,21 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import {
   Header,
   PostCard,
-  SuggestionsCard,
   ProfileCard,
   EditProfileModal,
+  Suggestions,
 } from "../../components/";
 
-import { getPosts, getUsers } from "../../redux/asyncThunk/";
-import { useSelector, useDispatch } from "react-redux";
+import { useSelector } from "react-redux";
 
 import "./Profile.css";
 
 const Profile = () => {
   const [showEditModal, setShowEditModal] = useState(false);
-  const dispatch = useDispatch();
-  const { users } = useSelector((state) => state.users);
+
   const { posts } = useSelector((state) => state.posts);
   const { user } = useSelector((state) => state.auth);
-
-  useEffect(() => {
-    dispatch(getUsers());
-    dispatch(getPosts());
-  }, [dispatch]);
-
-  const otherUsers = users.filter((item) => item.username !== user.username);
-
-  const nonFollowing = otherUsers.filter((item) =>
-    item.following.every((following) => following.username !== user.username)
-  );
 
   const userPosts = posts.filter((post) => post.username === user.username);
 
@@ -64,14 +51,7 @@ const Profile = () => {
             </section>
           </section>
 
-          <section className="suggestions-container">
-            <h2 className="suggestions-title">People you may know</h2>
-            <div className="suggestions-list">
-              {nonFollowing.map((user) => (
-                <SuggestionsCard key={user._id} user={user} />
-              ))}
-            </div>
-          </section>
+          <Suggestions />
         </section>
       </main>
     </>
